# backend/internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "weflow swagger api documention.",
        "title": "weflow",
        "contact": {
            "name": "hertz-contrib",
            "url": "https://github.com/hertz-contrib"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/weflow",
    "paths": {
...
        "/insttask/del": {
            "post": {
                "description": "删除实例任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实例任务"
                ],
                "summary": "删除实例任务",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "InstTaskDeleteVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.InstTaskDeleteVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/insttask/initiated": {
            "post": {
                "description": "获取发起的实例任务列表（已发起）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实例任务"
                ],
                "summary": "获取发起的实例任务列表（已发起）",
                "parameters": [
                    {
                        "description": "已发起的请求参数",
                        "name": "InstTaskQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.InstTaskQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.InstTaskResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/insttask/resume": {
            "post": {
                "description": "恢复实例任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实例任务"
                ],
                "summary": "恢复实例任务",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "InstTaskSesumeVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.InstTaskSesumeVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/insttask/start": {
            "post": {
                "description": "发起实例任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实例任务"
                ],
                "summary": "发起实例任务",
                "parameters": [
                    {
                        "description": "发起实例任务的请求参数",
                        "name": "InstTaskStartVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.InstTaskStartVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/insttask/stop": {
            "post": {
                "description": "终止实例任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实例任务"
                ],
                "summary": "终止实例任务",
                "parameters": [
                    {
                        "description": "终止实例任务的请求参数",
                        "name": "InstTaskStopVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.InstTaskStopVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/insttask/suspend": {
            "post": {
                "description": "挂起实例任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实例任务"
                ],
                "summary": "挂起实例任务",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "InstTaskSuspendVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.InstTaskSuspendVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "base.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "0:成功，其他：失败",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
...
        "bo.InstTaskResult": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string"
                },
                "createUserID": {
                    "description": "创建人id",
                    "type": "string"
                },
                "createUserName": {
                    "description": "创建人名称",
                    "type": "string"
                },
                "endTime": {
                    "description": "结束时间 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string"
                },
                "formDefID": {
                    "description": "表单定义id",
                    "type": "string"
                },
                "id": {
                    "description": "唯一id",
                    "type": "integer"
                },
                "instTaskID": {
                    "description": "实例任务id",
                    "type": "string"
                },
                "modelID": {
                    "description": "模板id",
                    "type": "string"
                },
                "processDefID": {
                    "description": "流程定义id",
                    "type": "string"
                },
                "remark": {
                    "description": "描述",
                    "type": "string"
                },
                "startTime": {
                    "description": "发起时间 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string"
                },
                "status": {
                    "description": "任务状态【1：创建中(草稿)；2：进行中； 3：终止； 4：完成； 5：挂起；6：回退】",
                    "type": "integer"
                },
                "taskName": {
                    "description": "实例任务名称",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string"
                },
                "updateUserID": {
                    "description": "更新人id",
                    "type": "string"
                },
                "updateUserName": {
                    "description": "更新人名称",
                    "type": "string"
                },
                "versionID": {
                    "description": "版本id",
                    "type": "string"
                }
            }
        },
...
        "vo.InstTaskDeleteVO": {
            "type": "object",
            "properties": {
                "instTaskID": {
                    "description": "实例任务ID",
                    "type": "string",
                    "example": ""
                },
                "opinionDesc": {
                    "description": "操作意见",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.InstTaskQueryVO": {
            "type": "object",
            "properties": {
                "createTimeEnd": {
                    "description": "提交审批时间-结束 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string",
                    "example": ""
                },
                "createTimeStart": {
                    "description": "提交审批时间-开始 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string",
                    "example": ""
                },
                "finishTimeEnd": {
                    "description": "完成审批时间-结束 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string",
                    "example": ""
                },
                "finishTimeStart": {
                    "description": "完成审批时间-开始 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string",
                    "example": ""
                },
                "instStatus": {
                    "description": "任务状态，零值为全部",
                    "type": "integer",
                    "example": 0
                },
                "modelID": {
                    "description": "模板ID",
                    "type": "string",
                    "example": ""
                },
                "pageNum": {
                    "description": "页码",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer",
                    "example": 30
                },
                "taskName": {
                    "description": "任务名称",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.InstTaskSesumeVO": {
            "type": "object",
            "properties": {
                "instTaskID": {
                    "description": "实例任务ID",
                    "type": "string",
                    "example": ""
                },
                "opinionDesc": {
                    "description": "操作意见",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.InstTaskStartVO": {
            "type": "object",
            "properties": {
                "modelID": {
                    "description": "模板ID",
                    "type": "string",
                    "example": ""
                },
                "params": {
                    "description": "参数",
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "vo.InstTaskStopVO": {
            "type": "object",
            "properties": {
                "instTaskID": {
                    "description": "实例任务ID",
                    "type": "string",
                    "example": ""
                },
                "opinionDesc": {
                    "description": "操作意见",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.InstTaskSuspendVO": {
            "type": "object",
            "properties": {
                "instTaskID": {
                    "description": "实例任务ID",
                    "type": "string",
                    "example": ""
                },
                "opinionDesc": {
                    "description": "操作意见",
                    "type": "string",
                    "example": ""
                }
            }
        },
...
    }
}